feat: ignoreCase en busqueda RFID y arreglos del modelo

+++ src/main/java/Laboratorio_lex/modules/sincronizacion/repository/ConfiguracionExportacionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ConfiguracionExportacionRepository extends JpaRepository<ConfiguracionExportacion, Short> {
+public interface ConfiguracionExportacionRepository extends JpaRepository<ConfiguracionExportacion, Integer> {
 
     // Fila única de configuración (sembrada en V2)
     Optional<ConfiguracionExportacion> findFirstByOrderByIdAsc();
